Run leapfrog model fit

Run leapfrog model fit

run_model(
  parameters,
  configuration = "HivFullAgeStratification",
  output_years = seq(1970, 2030)
)

Arguments

parameters

Projection parameters

configuration

The model configuration to run, see list_model_configurations() for available configurations

output_years

Which years of the model to return from the simulation, defaults to all years from 1970 to 2030. Also used to control what years the simulation is run for. If output only 2030, simulation will be run from projection_start_year passed in the parameters list.

Value

List of model outputs, where the last dimension of each element is time, e.g. p_total_pop state variable has dimensions 81 x 2. If output_years specified has length 61 then the p_total_pop output will have dimensions 81 x 2 x 61.

Examples

pjnz <- system.file(
  "pjnz/bwa_aim-adult-art-no-special-elig_v6.13_2022-04-18.PJNZ",
  package = "frogger", mustWork = TRUE)
parameters <- prepare_leapfrog_parameters(pjnz, use_coarse_age_groups = TRUE)
out <- run_model(parameters, "HivCoarseAgeStratification", 1970:2030)