Create a custom Quarto project

This function is a wrapper for quarto::quarto_create_project() and adds some other froggeR goodies. First, a new Quarto project is created in the provided name directory. Then, the original .qmd file created from the quarto::quarto_create_project() is replaced with froggeR::write_quarto().

Usage

quarto_project(name, base_dir = getwd(), default = TRUE)

Arguments

name

The name of the Quarto project directory and initial .qmd file.

base_dir

Base directory where the project should be created. Defaults to current working directory.

default

Set to TRUE. This will use the custom YAML within the initial .qmd file. Change to FALSE if you would like the initial Quarto document to have a standard YAML header. See ?froggeR::write_quarto() for more details on the default setting.

Value

A Quarto project with:
* .qmd
* .gitignore
* README.md
* DATED_PROGRESS_NOTES.md
* custom.scss
* .Rproj

Details

Since this will most likely be a new project directory, froggeR searches for the existence of these other files and creates them by default:
* _variables.yml
* .gitignore
* README.md
* dated_progress_notes.md
* custom.scss
* .Rproj

If these files do not exist, froggeR will write them to the new project directory.

Examples

if (FALSE) { # \dontrun{
# Create a new Quarto project that uses a custom formatted YAML header and
# all other listed files:
quarto_project('frog_project', base_dir = getwd(), default = TRUE)

# Create a new Quarto project with generic YAML and all other listed files:
quarto_project('frog_project_2', base_dir = getwd(), default = FALSE)
} # }