Start a Quarto file with a formatted template header

This function creates a new Quarto document with a pre-formatted YAML header. The default template requires a _variables.yml file. If none exists in the project, a prompt is provided to create the file. The _variables.yml allows for a reusable custom header that will be applied to all froggeR default-style Quarto documents.

Usage

write_quarto(filename = "new", path = getwd(), default = TRUE, proj = FALSE)

Arguments

filename

The name of the file. Do not include '.qmd'

path

The path to the main project level. Defaults to the current working directory.

default

The default is set to TRUE and will create a Quarto template file that pulls infomation from the folder/_variables.yml file.

proj

Set to TRUE for Quarto projects (for internal use).

Value

A Quarto document with formatted YAML and two blank starter sections.

Details

When default = TRUE, the project path is checked for a SCSS sheet. If none is found, a prompt offers to create one.

NOTE: Please visit https://quarto.org/docs/output-formats/html-themes.html to explore more Quarto theming options.

Examples

if (FALSE) { # \dontrun{
# To create new_doc.qmd:
write_quarto(filename = "new_doc", path = "path/to/project")
} # }